[Enhancement #145] Add generator of Manageable entity classes.
